+++ client/js/component/container/Signup.jsx
@@ -5,7 +5,7 @@ import signupStyles from '../../../sass/signup.scss';
 import TopNavigation from '../layout/TopNavigation.jsx';
 import Footer from '../layout/Footer.jsx';
 
-import { createUser } from '../../actions/userActions';
+import { createUser } from '../../actions/authAction';
 
 @connect((store) => {
   return  {
